% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_glassnode_api.R
\name{call_glassnode_api}
\alias{call_glassnode_api}
\title{Call Glassnode API}
\usage{
call_glassnode_api(path, params)
}
\arguments{
\item{path}{The path to pass to the API URL}

\item{params}{A list of parameters to pass to the API}
}
\value{

}
\description{
Call Glassnode API
}
