% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lightening_network.R
\name{call_lightning_api}
\alias{call_lightning_api}
\title{call the GN API for lightening parameters}
\usage{
call_lightning_api(var_name, params)
}
\arguments{
\item{var_name}{Specific lightning network parameter}

\item{params}{list of parameters to query the API with}
}
\value{
a tibble with the datetime and the specific attribute
}
\description{
call the GN API for lightening parameters
}
