% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_addresses_profit_loss.R
\name{get_addresses_profit}
\alias{get_addresses_profit}
\alias{get_addresses_loss}
\title{Get Addresses in Profit or Loss}
\usage{
get_addresses_profit(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_addresses_loss(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and address count
}
\description{
These functions return the number of addresses with unrealized profit or loss.
Addresses in profit have an average buy price lower than the current price,
while addresses in loss have an average buy price higher than the current price.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_addresses_profit()
y <- get_addresses_loss()
}
}
