% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block_data.R
\name{get_block_height}
\alias{get_block_height}
\alias{get_block_interval_mean}
\alias{get_block_interval_median}
\alias{get_block_size_mean}
\alias{get_blocks_count}
\title{Get Blockchain Block Metrics}
\usage{
get_block_height(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_block_interval_mean(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_block_interval_median(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_block_size_mean(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_blocks_count(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and block metrics
}
\description{
These functions retrieve fundamental blockchain metrics related to blocks,
including block intervals, sizes, and counts. These provide insights into
network performance and blockchain growth.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_block_height()
y <- get_block_interval()
z <- get_block_size()
}
}
