% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\name{get_bridge_deposits}
\alias{get_bridge_deposits}
\title{Get Bridge Deposits by Chain}
\usage{
get_bridge_deposits(
  asset = "ETH",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{Asset to examine (default: "ETH")}

\item{since}{Optional start date for data retrieval}

\item{until}{Optional end date for data retrieval}

\item{frequency}{Frequency of data points (default: "24h")}

\item{api_key}{API key for Glassnode (defaults to environment variable)}

\item{as_date}{Logical to return Date instead of POSIXct for daily frequency}
}
\value{
A tibble with date and bridge deposit data
}
\description{
Retrieves bridge deposit data for specified blockchain assets
}
\examples{
\dontrun{
x <- get_bridge_deposits("ETH")
}
}
