% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_btc_dominance.R
\name{get_btc_dominance}
\alias{get_btc_dominance}
\title{Get BTC Dominance}
\usage{
get_btc_dominance(
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and BTC dominance percentage
}
\description{
BTC Dominance represents the percentage of the total cryptocurrency market cap
that is represented by Bitcoin. It's a key metric for understanding Bitcoin's
relative position in the crypto market.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_btc_dominance()
}
}
