% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply.R
\name{get_burn_rate}
\alias{get_burn_rate}
\title{Get Burn Rate}
\usage{
get_burn_rate(
  asset = "ETH",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and burn rate
}
\description{
The rate at which supply is being burned per interval (relevant for
EIP-1559 assets like ETH).
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_burn_rate()
}
}
