% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_circulable_supply.R
\name{get_circulable_supply}
\alias{get_circulable_supply}
\title{Get the total amount of a coin that is availble at a given moment}
\usage{
get_circulable_supply(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{The asset whose supply to get (see list on Glassnode)}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
a tidy data frame of data
}
\description{
Get the total amount of a coin that is availble at a given moment
}
\examples{
\dontrun{
# Need a valid API key to run
x <- get_circulable_supply()
}
}
