% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closing_price.R
\name{get_closing_price}
\alias{get_closing_price}
\title{Get the closing price for a given token}
\usage{
get_closing_price(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}

\item{exchange}{What exchange to query. Defaults to all ("aggregated")}

\item{currency}{Return values in native tokens or in USD (the default)}
}
\value{
Returns a tibble with columns for the datetime and price
}
\description{
Get the closing price for a given token
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_closing_price()
}
}
