% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_balance.R
\name{get_exchange_balance}
\alias{get_exchange_balance}
\title{Get Balances for Exchange Wallets}
\usage{
get_exchange_balance(params = list(), api_key = Sys.getenv("API_KEY"))
}
\arguments{
\item{params}{A list of parameters to pass to the API call}

\item{api_key}{The API key to use. By default it will check the API_KEY evn variable}
}
\value{

}
\description{
Get Balances for Exchange Wallets
}
