% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_flows.R
\name{get_exchange_inflows}
\alias{get_exchange_inflows}
\alias{get_exchange_outflows}
\alias{get_exchange_netflows}
\alias{get_exchange_deposits_count}
\title{Get Exchange Flow Metrics}
\usage{
get_exchange_inflows(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_exchange_outflows(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_exchange_netflows(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_exchange_deposits_count(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and exchange flow metrics
}
\description{
These functions retrieve exchange inflow and outflow metrics, providing insights
into exchange activity and capital movements between exchanges and the broader network.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_exchange_inflows()
y <- get_exchange_outflows()
z <- get_exchange_netflows()
}
}
