% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fees.R
\name{get_fees_mean}
\alias{get_fees_mean}
\alias{get_fees_median}
\alias{get_fees_total}
\alias{get_exchange_fees}
\title{Get Network Fee Metrics}
\usage{
get_fees_mean(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_fees_median(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_fees_total(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_exchange_fees(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and fee metrics
}
\description{
These functions retrieve various fee-related metrics from the blockchain network.
They provide insights into transaction costs and network economics.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_fees_mean()
y <- get_fees_total()
z <- get_fees_median()
}
}
