% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{get_futures_liquidations}
\alias{get_futures_liquidations}
\alias{get_futures_liquidations_long}
\alias{get_futures_liquidations_short}
\title{Get Futures Liquidations}
\usage{
get_futures_liquidations(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_futures_liquidations_long(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_futures_liquidations_short(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and liquidation volume
}
\description{
The total USD volume of long and/or short futures positions liquidated
over the given interval. A key indicator of deleveraging events and
derivatives-market stress.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_futures_liquidations()
}
}
