% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{get_futures_open_interest}
\alias{get_futures_open_interest}
\alias{get_futures_open_interest_perpetual}
\alias{get_futures_open_interest_crypto_margin}
\alias{get_futures_open_interest_cash_margin}
\alias{get_futures_open_interest_cme}
\alias{get_futures_open_interest_mcap_ratio}
\title{Get Futures Open Interest}
\usage{
get_futures_open_interest(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_futures_open_interest_perpetual(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_futures_open_interest_crypto_margin(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_futures_open_interest_cash_margin(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_futures_open_interest_cme(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_futures_open_interest_mcap_ratio(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and open interest
}
\description{
These functions retrieve aggregate open interest across tracked futures
exchanges, broken out by margin type and venue. Open interest is a core
measure of aggregate leverage in the derivatives market.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_futures_open_interest()
}
}
