% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{get_hodl_wave}
\alias{get_hodl_wave}
\title{Return the HODL waves for BTC or ETH}
\usage{
get_hodl_wave(
  asset = "BTC",
  since = NULL,
  until = NULL,
  api_key = Sys.getenv("GN_API_KEY")
)
}
\arguments{
\item{asset}{Asset to examine (BTC or ETH)}

\item{since}{The start date}

\item{until}{The end date}

\item{api_key}{The api key. It looks in the env variable GN_API_KEY if not set}
}
\value{
a tibble in long format with columns date, duration, and amount (share of circulating supply)
}
\description{
Return the HODL waves for BTC or ETH
}
\examples{
\dontrun{
x <- get_hodl_wave()
}
}
