% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply.R
\name{get_illiquid_supply_change}
\alias{get_illiquid_supply_change}
\alias{get_liquid_supply_change}
\title{Get Illiquid/Liquid Supply Change}
\usage{
get_illiquid_supply_change(
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_liquid_supply_change(
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and supply change
}
\description{
The 30-day net change in illiquid or liquid supply, i.e. the flow rather
than the stock versions of \code{get_illiquid_supply()}/\code{get_liquid_supply()}.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_illiquid_supply_change()
}
}
