% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply.R
\name{get_issued_supply}
\alias{get_issued_supply}
\title{Get Newly Issued Supply}
\usage{
get_issued_supply(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and issued supply
}
\description{
The amount of new supply issued (e.g. via block rewards) over the given interval.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_issued_supply()
}
}
