% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lightning_network_metrics.R
\name{get_lightning_channel_size_median}
\alias{get_lightning_channel_size_median}
\title{Get Lightning Network Channel Size (Median)}
\usage{
get_lightning_channel_size_median(
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and median channel size in BTC
}
\description{
The median BTC size of public Lightning Network channels.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_lightning_channel_size_median()
}
}
