% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lightening_network.R
\name{get_lightning_network}
\alias{get_lightning_network}
\title{Get statistics on the Lightning Network}
\usage{
get_lightning_network(
  since = NULL,
  until = NULL,
  frequency = "24h",
  currency = "USD",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{since, until}{A POSIX compatible value that will be converted to a unix data number}

\item{frequency}{What resolution to use. See API docs for possible values. The default is "24h" (daily)}

\item{currency}{What currency to use. Choices are "USD" or "NATIVE"}

\item{api_key}{The api key used to retrieve the data}

\item{as_date}{A logical for turning the datatime object returned to a date object if it is daily data}
}
\value{
A tidy data set of characteristics
}
\description{
This function retrieves all available statistisc on the lightning network and returns them
as a tidy data set.
}
\details{
The specific functions it returns are:
 - Mean channel size (channel_size_mean)
 - Median channel size (channel_size_median)
 - Total capacity of the network (network_capacity_sum)
 - The total number of channels (channels_count)
 - The total number of nodes (nodes_count)
 The default is to return the value amounts as USD unless "NATIVE" is chosen for currency
}
\examples{
\dontrun{
# Need a valid API key to run
x <- get_lightning_network
}
}
