% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mempool.R
\name{get_mempool_count}
\alias{get_mempool_count}
\alias{get_mempool_size}
\alias{get_mempool_fees}
\alias{get_mempool_value}
\title{Get Mempool Metrics}
\usage{
get_mempool_count(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_mempool_size(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_mempool_fees(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_mempool_value(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and mempool metrics
}
\description{
These functions retrieve various mempool-related metrics showing transactions
waiting to be confirmed. They provide insights into network congestion and
transaction fee market dynamics.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_mempool_count()
y <- get_mempool_size()
z <- get_mempool_fees()
}
}
