% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mempool.R
\name{get_mempool_txs_size_distribution}
\alias{get_mempool_txs_size_distribution}
\title{Get Mempool Transaction Size Distribution by Relative Fee}
\usage{
get_mempool_txs_size_distribution(
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with a datetime column and one \code{txs_size_distribution_<bucket>} column per relative-fee cohort (e.g. \code{0_1}, \code{1_2}, ..., \code{50_plus})
}
\description{
The size of transactions waiting in different relative fee cohorts.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_mempool_txs_size_distribution()
}
}
