% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\name{get_miner_vol}
\alias{get_miner_vol}
\title{Title}
\usage{
get_miner_vol(
  miner = "aggregated",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{as_date}{}
}
\value{

}
\description{
Title
}
