% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mining_difficulty.R
\name{get_mining_difficulty}
\alias{get_mining_difficulty}
\title{Get Mining Difficulty}
\usage{
get_mining_difficulty(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and mining difficulty
}
\description{
The current estimated number of hashes required to mine a block.
Mining difficulty adjusts to maintain consistent block times as hash rate changes.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_mining_difficulty()
}
}
