% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply.R
\name{get_minted_supply}
\alias{get_minted_supply}
\title{Get Minted Supply}
\usage{
get_minted_supply(
  asset = "ETH",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. As of this metric's introduction, Glassnode only supports \code{"ETH"} for it}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and minted supply
}
\description{
The amount of new supply minted over the given interval (relevant for
smart-contract assets and stablecoins).
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_minted_supply()
}
}
