% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentage_top_balance.R
\name{get_percent_top_balance}
\alias{get_percent_top_balance}
\title{Get the percent of circulable supply for a crypto that is at deposited in
large accoints}
\usage{
get_percent_top_balance(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime, and percent
}
\description{
This function calls top 1% balance call. It returns a tibble with the amount
that is the percent (in decimal form) of the crypto in wallets containing
the largest top 1% of balances. Smart contracts, exchanges and other identified
wallets are excluded.
}
\details{
NOTE: FOr BTC this is entities (cluster of wallets) for others this is wallets
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_percent_top_balance()
}
}
