% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provably_lost_coins.R
\name{get_provably_lost_coins}
\alias{get_provably_lost_coins}
\title{Get the amount of coins that are considered lost or HODLed}
\usage{
get_provably_lost_coins(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{Asset to examine (BTC or LTC)}

\item{since}{The start date}

\item{until}{The end date}

\item{frequency}{The frequency to examine: 1h,24h}

\item{api_key}{The api key. It looks in the env variable GN_API_KEY if not set}

\item{as_date}{A logical to return a Date not a POSIXct object for daily or lower frequency}
}
\value{
return a tibble with the date and the addresses
}
\description{
This is a calculated statistics
}
\examples{
\dontrun{
x <- get_new_addresses()
}
}
