% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{get_spot_volume_mcap_ratio}
\alias{get_spot_volume_mcap_ratio}
\title{Get Spot Volume to Marketcap Ratio}
\usage{
get_spot_volume_mcap_ratio(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and spot volume/marketcap ratio
}
\description{
The ratio of daily spot trading volume to market capitalization, a measure of
relative liquidity/turnover.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_spot_volume_mcap_ratio()
}
}
