% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supply_contracts.R
\name{get_supply_contracts}
\alias{get_supply_contracts}
\title{Get amount of a given asset is deposited in smart contracts}
\usage{
get_supply_contracts(
  asset = "ETH",
  since = NULL,
  until = NULL,
  api_key = Sys.getenv("GN_API_KEY")
)
}
\arguments{
\item{asset}{The asset to query. Default is ETH}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}
}
\value{
a tidy data frame of data
}
\description{
This function returns the percentage (in decimal form) of all
circulable supply of the asset that is deposited in a smart contract
}
\examples{
\dontrun{
# Need a valid API key to run
x <- get_supply_contracts()
}
}
