% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supply_extended.R
\name{get_supply_in_loss}
\alias{get_supply_in_loss}
\title{Get Supply in Loss}
\usage{
get_supply_in_loss(
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and supply in loss
}
\description{
The amount of supply that is currently in loss.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_supply_in_loss()
}
}
