% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{get_total_addresses}
\alias{get_total_addresses}
\title{Get the number of unique addresses that was ever active}
\usage{
get_total_addresses(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{Asset to examine (see glassnode api docs for a list)}

\item{since}{The start date}

\item{until}{The end date}

\item{frequency}{The frequency to examine: 1h,24h,10m,1w,1month}

\item{api_key}{The api key. It looks in the env variable GN_API_KEY if not set}

\item{as_date}{A logical to return a Date not a POSIXct object for daily or lower frequency}
}
\value{
return a tibble with the date and the addresses
}
\description{
Get the number of unique addresses that was ever active
}
\examples{
\dontrun{
x <- get_total_addresses()
}
}
