% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transaction_volumes.R
\name{get_volume_adjusted_mean}
\alias{get_volume_adjusted_mean}
\alias{get_volume_adjusted_median}
\alias{get_volume_adjusted_total}
\alias{get_volume_entity_adjusted}
\title{Get Transaction Volume Metrics}
\usage{
get_volume_adjusted_mean(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_volume_adjusted_median(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_volume_adjusted_total(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_volume_entity_adjusted(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and volume metrics
}
\description{
These functions retrieve various transaction volume metrics including
change-adjusted volumes and entity-adjusted volumes that provide insights
into economic activity on the blockchain.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_volume_adjusted_mean()
y <- get_volume_adjusted_total()
z <- get_volume_entity_adjusted()
}
}
