% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transactions_rate.R
\name{get_transactions_rate}
\alias{get_transactions_rate}
\title{Get the total rate of transactions per second in the given frequency}
\usage{
get_transactions_rate(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{The chain to examine. Limited to BTC, ETH, LTC}

\item{since}{The start date}

\item{until}{The end date}

\item{frequency}{The frequency to examine: 1h,24h,10m,1w,1month}

\item{api_key}{The api key. It looks in the env variable GN_API_KEY if not set}

\item{as_date}{A logical to return a Date not a POSIXct object for daily or lower frequency}
}
\value{
This function returns a tibble consisting of the date and transaction rates.
}
\description{
Get the total rate of transactions per second in the given frequency
}
\examples{
\dontrun{
x <- get_transactions_rate()
}
}
