% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treasuries.R
\name{get_treasury_flows_net_companies}
\alias{get_treasury_flows_net_companies}
\alias{get_treasury_flows_net_governments}
\title{Get Net Treasury Flows}
\usage{
get_treasury_flows_net_companies(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_treasury_flows_net_governments(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the GN_API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with a datetime column and one net-flow column per tracked entity (prefixed with the function's metric name)
}
\description{
The net daily change in the amount of the asset held by corporate or
government treasuries.
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_treasury_flows_net_companies()
}
}
