% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tvl.R
\name{get_tvl}
\alias{get_tvl}
\title{Get the Total Value Locked for all DeFi platforms}
\usage{
get_tvl(params = list(), api_key = Sys.getenv("API_KEY"))
}
\arguments{
\item{params}{A list of parameters to pass to the API}

\item{api_key}{The API key to use. Checks the env variable API_KEY}
}
\value{

}
\description{
Get the Total Value Locked for all DeFi platforms
}
