% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{get_utxo_count}
\alias{get_utxo_count}
\alias{get_utxo_created_count}
\alias{get_utxo_spent_count}
\alias{get_utxo_created_value_mean}
\alias{get_utxo_created_value_median}
\alias{get_utxo_created_value_sum}
\alias{get_utxo_spent_value_mean}
\alias{get_utxo_spent_value_median}
\alias{get_utxo_spent_value_sum}
\title{Get UTXO Set Metrics}
\usage{
get_utxo_count(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_utxo_created_count(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_utxo_spent_count(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_utxo_created_value_mean(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_utxo_created_value_median(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_utxo_created_value_sum(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_utxo_spent_value_mean(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_utxo_spent_value_median(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_utxo_spent_value_sum(
  asset = "BTC",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{asset}{This is the asset to get information on. The list of available assets is on the GN API site}

\item{since, until}{A POSIX compatible date-time object. It's converted to a unix date number}

\item{frequency}{A resolution for the data. See API documentation but it defaults to 24h}

\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{as_date}{A logical to return a date-time object or a date object for daily observations}
}
\value{
Returns a tibble with columns for the datetime and UTXO metric
}
\description{
These functions retrieve metrics describing the Unspent Transaction Output
(UTXO) set: how many UTXOs exist, how many are created/spent per interval,
and the value created/spent. These describe the raw economic throughput of
UTXO-based blockchains (e.g. Bitcoin).
}
\examples{
\dontrun{
#Need a valid API key to run
x <- get_utxo_count()
}
}
