% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_params.R
\name{make_params}
\alias{make_params}
\title{Regularize parameters into the Glassnode API}
\usage{
make_params(api_key = Sys.getenv("GN_API_KEY"), ...)
}
\arguments{
\item{api_key}{The API key to use. By default it will check the API_KEY environmental variable}

\item{...}{A list of parameters for use}
}
\value{
A list of parameters ready to send to the GN API
}
\description{
This function makes sure some of the basic parameters are there for
the Glassnode API. Also ensures a return format that is known. Also defaults
the frequency to daily unless overridden
}
\details{
TODO: add more error checking to this function
}
