% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_date.R
\name{sanitize_date}
\alias{sanitize_date}
\title{Turn dates to unix datetime string}
\usage{
sanitize_date(the_date)
}
\arguments{
\item{the_date}{}
}
\value{
a unix numeric for datetime
}
\description{
This function is a utility function for make_params to make a number to
work in the API
}
\examples{
x <- sanitize_date("2020-01-01")
}
