% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_glassnode_api.R
\name{call_glassnode_api}
\alias{call_glassnode_api}
\title{Call Glassnode API}
\usage{
call_glassnode_api(path, params)
}
\arguments{
\item{path}{The path to pass to the API URL}

\item{params}{a set of parameters to pass the API.}
}
\value{
a parsed list from the JSON structure
}
\description{
This is the core function connecting to Glassnode. Since all (most?) API
 calls have the same seven parameters it does some error checking
 and adds defaults all the
}
\examples{
\dontrun{
# Need a valid API to run
x <- call_glassnode_api()
}
}
