% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hodl_wave.R
\name{get_hodl_wave}
\alias{get_hodl_wave}
\title{Return the HODL waves for BTC or ETH}
\usage{
get_hodl_wave(
  asset = "BTC",
  since = NULL,
  until = NULL,
  api_key = Sys.getenv("GN_API_KEY")
)
}
\arguments{
\item{api_key}{}
}
\value{
a Tibble with the waves as columns
}
\description{
Return the HODL waves for BTC or ETH
}
\examples{
x <- get_hodl_wave()
}
