% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining.R
\name{get_vol_miners}
\alias{get_vol_miners}
\alias{get_miner_vol}
\title{Get volume mined  for BTC miners}
\usage{
get_vol_miners(
  since = NULL,
  until = NULL,
  frequency = "24h",
  get_agg = TRUE,
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)

get_miner_vol(
  miner = "aggregated",
  since = NULL,
  until = NULL,
  frequency = "24h",
  api_key = Sys.getenv("GN_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{since}{The date to start the extraction}

\item{until}{The date to end the extraction}

\item{frequency}{What frequency to pull (either "1h" or "24h")}

\item{get_agg}{A logical for whether to get all miners}

\item{api_key}{The API key. It defaults to the GN_API_KEY environmental variables}

\item{as_date}{A logical whether to return a date object if the frequency is "24h"}

\item{miner}{The name of the miner to use}
}
\description{
These functions grab the number of coins  miners have won.
 The function get_vol_miners
returns either jsut the aggregated (if get_agg is true) or all miner
volumes. Individual miner's volumescan be found using get_miner_vol and
passing the name of the miner.
}
\details{
NOTE: These amounts are available only for BTC.
}
