% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_params}
\alias{make_params}
\title{Prepare parameters for Glassnode API}
\usage{
make_params(api_key = Sys.getenv("GN_API_KEY"), ...)
}
\arguments{
\item{api_key}{API key (defaults to environment variable)}

\item{...}{Additional parameters for the API call}
}
\value{
A named list of sanitized API parameters
}
\description{
Sanitizes and standardizes parameters for Glassnode API calls
}
